'Pre-choose' events with no alternative available times. Refactor-corrected spelling of 'compulsory' through all source code.
Corrected spelling of 'compulsory' in all files.

Events which have no alternatives on the timetable (eg. unique lectures)
will be 'pre-chosen.' That is, users don't need to press "choose" on the
event as this will be done for them if there are no other alternative
timeslots for the same item.

This is implemented by adding a .solo tag to each item, which is set to
true if it is the only event of its type in the course. Its choose tag
is then artificially clicked after adding it to the calendar.

(Note: I am unsure what all of the 'compulsory' things throughout the
code are actually meant to refer to or whether they were meant to refer
to classes with no alternative as I addressed in my other changes. Or is
it that some courses have optional events as well as a range of
alternative compulsory ones? It didn't seem to apply to anything in any
of my courses, or I grossly misunderstood it. I'm new to the uni and a
pleb.)